status: show score board
The score is calculated based on the number of clicks the user has done,
the elapsed time, number of mines and the size of the board. In this way
there will not be a situation where with the same setting you would get
higher score throught the castom setting than with the predefined
difficulty levels.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>