minesweeper: use pqueue as the record ceeper
The records are now sorted in priority queue. The queue has dynamic
sorting applied to it. This allows us to be able to create sorters and
change the sorting in settings.

We also applied the hash to the structure. In this way if a user with
the same name does the same socre that is already in the table, we will
be able to find out that they are not the same.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>