minesweaper: save pointer to the app in the layer
This allows any layer of the application to access any other layer. The
limitation is, that the layer must be already loaded. Therefore, we use
assert to check for this edge case.

The application must be used with weak_ptr. Otherwise we would have
cyclic dependency. The destructors would not be called and we could have
memory leaks.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>