tile: code readability
The function blongsToOcupant is changed to belongsTo. The argument
player is not a player it is the number (flag) that determines the state
of the tile. Therefore, the player is changed to ocupant.

At the same time, the belongsToOponent can be removed. It is not used
anywhere in the code.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>